refactor: move expo to Canvas, allow custom renderers
